fix(consistency): task-4186 money_float 래칫 위반 해소 (13->8, baseline 10)
check_consistency.py의 money_float 래칫이 10->13으로 늘어 CI가 지속
적색이었다(esc-health-ci_red, esc-ci-00f2dbd1410c 등). mandates 도메인
(min_cash_buffer_pct/cash_buffer_pct)과 market_data krx_data 어댑터
(price 3곳)의 화폐성 필드를 float에서 Decimal로 바꿨다. DB 행 매핑
(postgres_repository.py, resolve_binding.py)도 Decimal(str(...))로
맞췄다.

src/foundation/mandates/contracts/v1.py, src/foundation/risk/contracts/v1.py
의 동일 필드는 architecture guard P5.contract_type_changed(공개 v1 계약의
in-place 타입 변경 veto, 107번 계약 버전 표준)에 막혀 그대로 float로
남겼다 — application 계층(create_draft_mandate.py, propose_amendment.py,
evaluate_policy.py)에서 contract(float) <-> domain(Decimal) 경계를
명시적으로 변환한다. 이 두 계약 파일의 잔여 8건은 baseline(10) 이내라
DoD(money_float 0증가, <=10)를 만족한다 — v1 계약 자체의 Decimal 전환은
v2 계약 발행이 필요한 별도 스콥.

src/core/portfolio/mandate_binding.py(FROZEN_PAPER_ONLY)는 이미
Decimal(str(...)) 방어적 변환을 쓰고 있어 수정 없이 호환된다.

task-4186, L4 해당 없음(consistency 게이트 회귀 수정).

+++ src/foundation/mandates/domain/models.py
@@ -2,10 +2,12 @@
 
 Spec: AIOSproject 75_portfolio_mandate_l3_build_and_operational_specification_v1.0.md §1.
 """
+
 from __future__ import annotations
 
 from dataclasses import dataclass, field
 from datetime import datetime
+from decimal import Decimal
 from enum import Enum
 from uuid import UUID
 
@@ -42,7 +44,7 @@ class MandateRevision:
     state: MandateRevisionState
     max_total_exposure_pct: float
     max_single_instrument_pct: float
-    min_cash_buffer_pct: float
+    min_cash_buffer_pct: Decimal
     max_daily_loss_pct: float
     allowed_autonomy: Autonomy
     forbidden_assets: tuple[str, ...] = field(default_factory=tuple)
@@ -104,7 +106,7 @@ class PolicyEvaluationSubject:
     command_type: str
     instrument_exposure_pct: float | None = None
     total_exposure_pct: float | None = None
-    cash_buffer_pct: float | None = None
+    cash_buffer_pct: Decimal | None = None
     projected_daily_loss_pct: float | None = None
     requested_autonomy: Autonomy | None = None
     asset: str | None = None
